+++ desktop-widgets/command_private.h
@@ -1,12 +1,11 @@
 // SPDX-License-Identifier: GPL-2.0
 // Private definitions for the command-objects
 
+#include "core/dive.h"
+
 #include <vector>
 #include <utility>
 #include <QVector>
-
-struct dive;
-struct dive_trip;
 
 namespace Command {
 
@@ -52,5 +51,13 @@ void processByTrip(Vector &divesIn, Function action)
 	processByTrip(dives, action);
 
 }
+
+// Reset the selection to the dives of the "selection" vector and send the appropriate signals.
+// Set the current dive to "currentDive". "currentDive" must be an element of "selection" (or
+// null if "seletion" is empty). Return true if the selection or current dive changed.
+bool setSelection(const std::vector<dive *> &selection, dive *currentDive);
+
+// Get currently selectd dives
+std::vector<dive *> getDiveSelection();
 
 } // namespace Command
